Time: 3 ms (61.89%) | Memory: 23.6 MB (34.17%) - LeetSync

+++ 155-min-stack/min-stack.cpp
@@ -22,7 +22,7 @@ long long int minval ;
     
     void pop() {
         if(s.top() < minval ){
-            minval = (long long int )2*minval - s.top() ;
+            minval = 2*minval - s.top() ;
 
         }
         s.pop();
